requirements.txt
imutils==0.5.4
matplotlib==3.3.4
numpy==1.19.5
opencv_python_headless==4.5.4.60
Pillow==9.3.0
PyAutoGUI==0.9.53
scikit_learn==1.1.3
tensorflow==2.6.0
